+++ graphLayout/jetbrains.mps.graphLayout.graph/source_gen/jetbrains/mps/graphLayout/graph/ClusteredGraph.java
@@ -91,6 +91,6 @@ public class ClusteredGraph extends Graph implements IClusteredGraph {
   }
 
   public boolean isLeafCluster(Node cluster) {
-    return (int) ListSequence.fromList(cluster.getOutEdges()).count() == 0;
+    return ListSequence.fromList(cluster.getOutEdges()).count() == 0;
   }
 }
